Document new module layout
