+++ tests/test_entities.py
@@ -3,7 +3,7 @@ from types import SimpleNamespace
 
 import pytest
 
-import src.main as main
+import src.telegram_utils as tgu
 
 
 @pytest.mark.asyncio
@@ -15,13 +15,13 @@ async def test_get_chat_name_with_cache_and_client(monkeypatch):
             calls.append(ident)
             return SimpleNamespace(title="Chat Name")
 
-    main.client = DummyClient()
-    main.entity_name_cache.clear()
+    tgu.client = DummyClient()
+    tgu.entity_name_cache.clear()
 
-    name = await main.get_chat_name("id1", safe=True)
+    name = await tgu.get_chat_name("id1", safe=True)
     assert name == "Chat_Name"
     # Second call should hit cache and not call client again
-    name2 = await main.get_chat_name("id1", safe=True)
+    name2 = await tgu.get_chat_name("id1", safe=True)
     assert name2 == "Chat_Name"
     assert calls == ["id1"]
 
@@ -32,10 +32,10 @@ async def test_get_chat_name_error(monkeypatch):
         async def get_entity(self, ident):
             raise RuntimeError("fail")
 
-    main.client = FailClient()
-    main.entity_name_cache.clear()
+    tgu.client = FailClient()
+    tgu.entity_name_cache.clear()
 
-    name = await main.get_chat_name("https://t.me/testchat?param=1", safe=True)
+    name = await tgu.get_chat_name("https://t.me/testchat?param=1", safe=True)
     assert name == "testchat"
 
 
@@ -53,15 +53,15 @@ async def test_get_chat_name_various(monkeypatch, entity, expected):
         async def get_entity(self, ident):
             return entity
 
-    main.client = DummyClient()
-    main.entity_name_cache.clear()
-    result = await main.get_chat_name("identifier", safe=True)
+    tgu.client = DummyClient()
+    tgu.entity_name_cache.clear()
+    result = await tgu.get_chat_name("identifier", safe=True)
     assert result == expected
 
 
 @pytest.mark.asyncio
 async def test_get_chat_name_empty_identifier(monkeypatch):
-    result = await main.get_chat_name("", safe=True)
+    result = await tgu.get_chat_name("", safe=True)
     assert result == "chat_history"
 
 
@@ -71,9 +71,9 @@ async def test_get_chat_name_plus_link(monkeypatch):
         async def get_entity(self, ident):
             raise ValueError("not found")
 
-    main.client = DummyClient()
-    main.entity_name_cache.clear()
-    result = await main.get_chat_name("https://t.me/+abcDEF123", safe=True)
+    tgu.client = DummyClient()
+    tgu.entity_name_cache.clear()
+    result = await tgu.get_chat_name("https://t.me/+abcDEF123", safe=True)
     assert result == "invite_abcDEF123"
 
 
@@ -88,10 +88,10 @@ async def test_resolve_entities(monkeypatch):
                 raise RuntimeError("fail")
             return SimpleNamespace(id=int(ent))
 
-    main.client = DummyClient()
-    monkeypatch.setattr(main, "get_peer_id", lambda e: e.id)
+    tgu.client = DummyClient()
+    monkeypatch.setattr(tgu, "get_peer_id", lambda e: e.id)
 
-    result = await main.resolve_entities(["1", "bad", "2", "1"])
+    result = await tgu.resolve_entities(["1", "bad", "2", "1"])
     assert result == {1, 2}
     assert calls == ["1", "bad", "2", "1"]
 
@@ -105,11 +105,11 @@ async def test_get_entity_name_from_int(monkeypatch):
             recorded.append(type(ident))
             return SimpleNamespace(title="Chat")
 
-    main.client = DummyClient()
-    name = await main.get_entity_name(-1000000000042, safe=True)
+    tgu.client = DummyClient()
+    name = await tgu.get_entity_name(-1000000000042, safe=True)
     assert name == "Chat"
-    assert recorded and issubclass(recorded[0], main.types.PeerChannel)
+    assert recorded and issubclass(recorded[0], tgu.types.PeerChannel)
 
 
 def test_get_safe_name():
-    assert main.get_safe_name("A B") == "A_B"
+    assert tgu.get_safe_name("A B") == "A_B"
