+++ tests/test_main_flow.py
@@ -5,7 +5,11 @@ from types import SimpleNamespace
 
 import pytest
 
-import src.main as main
+import src.app as app
+import src.config as config_module
+import src.prompts as prompts
+import src.stats as stats_module
+import src.telegram_utils as tgu
 
 
 class BreakLoop(Exception):
@@ -29,12 +33,12 @@ async def test_rescan_loop(monkeypatch):
         return {}
 
     monkeypatch.setattr(asyncio, "sleep", fake_sleep)
-    monkeypatch.setattr(main, "update_instance_chat_ids", fake_update)
-    monkeypatch.setattr(main, "load_config", fake_load_config)
+    monkeypatch.setattr(app, "update_instance_chat_ids", fake_update)
+    monkeypatch.setattr(app, "load_config", fake_load_config)
 
-    inst = main.Instance(name="i", words=[], target_chat=0)
+    inst = app.Instance(name="i", words=[], target_chat=0)
     with pytest.raises(BreakLoop):
-        await main.rescan_loop(inst, interval=0)
+        await app.rescan_loop(inst, interval=0)
     assert sleep_calls == [0]
     assert len(load_calls) == 1
 
@@ -49,7 +53,7 @@ async def test_setup_logging(monkeypatch):
     monkeypatch.setattr(logging, "basicConfig", fake_basicConfig)
     tele_logger = logging.getLogger("telethon")
     tele_logger.setLevel(logging.INFO)
-    main.setup_logging("debug")
+    app.setup_logging("debug")
     assert recorded["level"] == logging.DEBUG
     assert tele_logger.level == logging.WARNING
 
@@ -57,46 +61,46 @@ async def test_setup_logging(monkeypatch):
 @pytest.mark.asyncio
 async def test_main_flow(monkeypatch, dummy_tg_client, dummy_message_cls, tmp_path):
     config = {"log_level": "info"}
-    monkeypatch.setattr(main, "load_config", lambda: config)
-    monkeypatch.setattr(main, "get_api_credentials", lambda cfg: (1, "h", "s"))
+    monkeypatch.setattr(app, "load_config", lambda: config)
+    monkeypatch.setattr(app, "get_api_credentials", lambda cfg: (1, "h", "s"))
 
     dummy_client = dummy_tg_client
-    monkeypatch.setattr(main, "TelegramClient", lambda s, a, b: dummy_client)
+    monkeypatch.setattr(app, "TelegramClient", lambda s, a, b: dummy_client)
 
     stats_path = tmp_path / "stats.json"
     monkeypatch.setattr(
-        main, "stats", main.StatsTracker(str(stats_path), flush_interval=0)
+        app, "stats", stats_module.StatsTracker(str(stats_path), flush_interval=0)
     )
 
     async def fake_rescan(inst):
         return None
 
-    monkeypatch.setattr(main, "rescan_loop", fake_rescan)
+    monkeypatch.setattr(app, "rescan_loop", fake_rescan)
 
     async def fake_update(inst, fr):
         inst.chat_ids = {1}
 
-    monkeypatch.setattr(main, "update_instance_chat_ids", fake_update)
+    monkeypatch.setattr(app, "update_instance_chat_ids", fake_update)
 
     async def fake_load_instances(cfg):
         return [
-            main.Instance(name="i", words=["hi"], target_chat=99, target_entity="name")
+            app.Instance(name="i", words=["hi"], target_chat=99, target_entity="name")
         ]
 
-    monkeypatch.setattr(main, "load_instances", fake_load_instances)
+    monkeypatch.setattr(app, "load_instances", fake_load_instances)
 
     async def fake_get_message_source(m):
         return "URL"
 
-    monkeypatch.setattr(main, "get_message_source", fake_get_message_source)
+    monkeypatch.setattr(tgu, "get_message_source", fake_get_message_source)
 
     async def fake_get_chat_name(v, safe=False):
         return "name"
 
-    monkeypatch.setattr(main, "get_chat_name", fake_get_chat_name)
+    monkeypatch.setattr(tgu, "get_chat_name", fake_get_chat_name)
 
-    await main.main()
-    assert main.config is config
+    await app.main()
+    assert app.config is config
 
     handler = dummy_client.on_handler
     msg = dummy_message_cls(SimpleNamespace(channel_id=1), msg_id=5, text="hi there")
@@ -120,20 +124,22 @@ async def test_process_message_prompt(monkeypatch, dummy_message_cls, tmp_path):
         async def send_message(self, *a, **k):
             sent.append((a, k))
 
-    main.client = DummyClient()
-    main.stats = main.StatsTracker(str(tmp_path / "stats.json"), flush_interval=0)
+    app.client = DummyClient()
+    app.stats = stats_module.StatsTracker(
+        str(tmp_path / "stats.json"), flush_interval=0
+    )
 
-    inst = main.Instance(
+    inst = app.Instance(
         name="p",
         words=[],
-        prompts=[main.Prompt(name="hi", prompt="hi", threshold=4)],
+        prompts=[prompts.Prompt(name="hi", prompt="hi", threshold=4)],
         target_chat=1,
     )
 
     async def fake_match(prompt, text, inst_name):
         assert prompt.prompt == "hi"
         assert inst_name == "p"
-        return main.EvaluateResult(similarity=5, main_fragment="")
+        return prompts.EvaluateResult(similarity=5, main_fragment="")
 
     async def fake_get_message_source(msg):
         return "src"
@@ -141,13 +147,13 @@ async def test_process_message_prompt(monkeypatch, dummy_message_cls, tmp_path):
     async def fake_get_chat_name(v, safe=False):
         return "n"
 
-    monkeypatch.setattr(main, "match_prompt", fake_match)
-    monkeypatch.setattr(main, "get_message_source", fake_get_message_source)
-    monkeypatch.setattr(main, "get_chat_name", fake_get_chat_name)
+    monkeypatch.setattr(app, "match_prompt", fake_match)
+    monkeypatch.setattr(tgu, "get_message_source", fake_get_message_source)
+    monkeypatch.setattr(tgu, "get_chat_name", fake_get_chat_name)
 
     msg = dummy_message_cls(SimpleNamespace(channel_id=1), msg_id=7, text="hi")
     event = SimpleNamespace(message=msg, chat_id=1)
-    await main.process_message(inst, event)
+    await app.process_message(inst, event)
 
     assert sent[0][0][0] == 1
     assert msg.forwarded == [1]
@@ -158,43 +164,43 @@ async def test_ignore_usernames(
     monkeypatch, dummy_tg_client, dummy_message_cls, tmp_path
 ):
     config = {"log_level": "info", "ignore_usernames": ["bad"]}
-    monkeypatch.setattr(main, "load_config", lambda: config)
-    monkeypatch.setattr(main, "get_api_credentials", lambda cfg: (1, "h", "s"))
+    monkeypatch.setattr(app, "load_config", lambda: config)
+    monkeypatch.setattr(app, "get_api_credentials", lambda cfg: (1, "h", "s"))
 
     dummy_client = dummy_tg_client
-    monkeypatch.setattr(main, "TelegramClient", lambda s, a, b: dummy_client)
+    monkeypatch.setattr(app, "TelegramClient", lambda s, a, b: dummy_client)
 
     stats_path = tmp_path / "stats.json"
     monkeypatch.setattr(
-        main, "stats", main.StatsTracker(str(stats_path), flush_interval=0)
+        app, "stats", stats_module.StatsTracker(str(stats_path), flush_interval=0)
     )
 
     async def fake_rescan(inst):
         return None
 
-    monkeypatch.setattr(main, "rescan_loop", fake_rescan)
+    monkeypatch.setattr(app, "rescan_loop", fake_rescan)
 
     async def fake_update(inst, fr):
         inst.chat_ids = {1}
 
-    monkeypatch.setattr(main, "update_instance_chat_ids", fake_update)
+    monkeypatch.setattr(app, "update_instance_chat_ids", fake_update)
 
     async def fake_load_instances(cfg):
-        return [main.Instance(name="i", words=["hi"], target_chat=99)]
+        return [app.Instance(name="i", words=["hi"], target_chat=99)]
 
-    monkeypatch.setattr(main, "load_instances", fake_load_instances)
+    monkeypatch.setattr(app, "load_instances", fake_load_instances)
 
     async def fake_get_message_source(m):
         return "URL"
 
-    monkeypatch.setattr(main, "get_message_source", fake_get_message_source)
+    monkeypatch.setattr(tgu, "get_message_source", fake_get_message_source)
 
     async def fake_get_chat_name(v, safe=False):
         return "name"
 
-    monkeypatch.setattr(main, "get_chat_name", fake_get_chat_name)
+    monkeypatch.setattr(tgu, "get_chat_name", fake_get_chat_name)
 
-    await main.main()
+    await app.main()
 
     handler = dummy_client.on_handler
     msg = dummy_message_cls(SimpleNamespace(channel_id=1), msg_id=5, text="hi")
@@ -203,51 +209,51 @@ async def test_ignore_usernames(
     await handler(event)
     assert msg.forwarded == []
     assert dummy_client.sent == []
-    assert main.stats.data["total"] == 0
+    assert app.stats.data["total"] == 0
 
 
 @pytest.mark.asyncio
 async def test_ignore_words(monkeypatch, dummy_tg_client, dummy_message_cls, tmp_path):
     config = {"log_level": "info"}
-    monkeypatch.setattr(main, "load_config", lambda: config)
-    monkeypatch.setattr(main, "get_api_credentials", lambda cfg: (1, "h", "s"))
+    monkeypatch.setattr(app, "load_config", lambda: config)
+    monkeypatch.setattr(app, "get_api_credentials", lambda cfg: (1, "h", "s"))
 
     dummy_client = dummy_tg_client
-    monkeypatch.setattr(main, "TelegramClient", lambda s, a, b: dummy_client)
+    monkeypatch.setattr(app, "TelegramClient", lambda s, a, b: dummy_client)
 
     stats_path = tmp_path / "stats.json"
     monkeypatch.setattr(
-        main, "stats", main.StatsTracker(str(stats_path), flush_interval=0)
+        app, "stats", stats_module.StatsTracker(str(stats_path), flush_interval=0)
     )
 
     async def fake_rescan(inst):
         return None
 
-    monkeypatch.setattr(main, "rescan_loop", fake_rescan)
+    monkeypatch.setattr(app, "rescan_loop", fake_rescan)
 
     async def fake_update(inst, fr):
         inst.chat_ids = {1}
 
-    monkeypatch.setattr(main, "update_instance_chat_ids", fake_update)
+    monkeypatch.setattr(app, "update_instance_chat_ids", fake_update)
 
     async def fake_load_instances(cfg):
         return [
-            main.Instance(name="i", words=["hi"], ignore_words=["bad"], target_chat=99)
+            app.Instance(name="i", words=["hi"], ignore_words=["bad"], target_chat=99)
         ]
 
-    monkeypatch.setattr(main, "load_instances", fake_load_instances)
+    monkeypatch.setattr(app, "load_instances", fake_load_instances)
 
     async def fake_get_message_source(m):
         return "URL"
 
-    monkeypatch.setattr(main, "get_message_source", fake_get_message_source)
+    monkeypatch.setattr(tgu, "get_message_source", fake_get_message_source)
 
     async def fake_get_chat_name(v, safe=False):
         return "name"
 
-    monkeypatch.setattr(main, "get_chat_name", fake_get_chat_name)
+    monkeypatch.setattr(tgu, "get_chat_name", fake_get_chat_name)
 
-    await main.main()
+    await app.main()
 
     handler = dummy_client.on_handler
     msg = dummy_message_cls(SimpleNamespace(channel_id=1), msg_id=5, text="bad hi")
@@ -255,4 +261,4 @@ async def test_ignore_words(monkeypatch, dummy_tg_client, dummy_message_cls, tmp
     await handler(event)
     assert msg.forwarded == []
     assert dummy_client.sent == []
-    assert main.stats.data["total"] == 0
+    assert app.stats.data["total"] == 0
